[INTERNAL] Adapters: Moved tests into new structure after rebase
